requirements.txt
streamlit>=1.20.0
google-genai
diffusers>=0.14.0
torch>=1.13.0
transformers>=4.28.0
Pillow>=9.0.0
python-dotenv
